Features Renamed Controller into GameController and changed implementation
